Cadastro e Remoção de usuário implementado
