+++ cam_settings.json
@@ -1,7 +1,7 @@
 {
     "camera_index" : 0,
     "fps" : 30,
-    "width" : 480,
-    "height" : 640,
+    "width" : 1080,
+    "height" : 1920,
     "camera_object" : null
 }
