--- classes_refact/BancoAlunos.py
@@ -1,5 +1,0 @@
-class BancoAlunos:
-    def __init__(self, path):
-        # dataframe de um excel
-        # id | nome
-        pass
