+++ src/BancoAlunos.py
@@ -1,0 +1,5 @@
+class BancoAlunos:
+    def __init__(self, path):
+        # dataframe de um excel
+        # id | nome
+        pass
