+++ src/liveness_detection_src/requirements.txt
@@ -1,0 +1,8 @@
+easydict==1.10
+numpy==1.23.5
+tqdm==4.64.1
+torchvision
+torch
+opencv-python==4.6.0.66
+Pillow==9.3.0
+tensorboardX==2.5.1
