+++ test.py
@@ -1,0 +1,3 @@
+import src
+
+src.iniciar()
